requirements-dev.txt
asgiref==3.11.0
Django==5.2.9
django-debug-toolbar==6.1.0
django-dotenv==1.4.2
django-reversion==6.1.0
django-widget-tweaks==1.5.1
dotenv==0.9.9
flake8==7.3.0
git-filter-repo==2.47.0
gunicorn==23.0.0
iniconfig==2.3.0
mccabe==0.7.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.2.1
ruff==0.14.10
sqlparse==0.5.5
whitenoise==6.11.0
